+++ client/tws_foundation_client/lib/src/core/extensions/string.dart
@@ -1,15 +1,15 @@
-extension DoubleExtension on double{
+extension StringExtension on String{
 
   /// Validate if a decimal value pass the number length paramaters.
   /// Returns TRUE when the double value is valid, else, return FALSE.
-  bool validateDecimal(String value, {int maxIntegers = 4, int maxDecimals = 6}){
+  bool evaluateAsDecimal({int maxIntegers = 4, int maxDecimals = 6}){
     //skip default values.
-    if(value == "0" || value == "0.0") return true;
-    double? parsedValue = double.tryParse(value);
+    if(this == "0" || this == "0.0") return true;
+    double? parsedValue = double.tryParse(this);
     //validate if the text input actually is a decimal value.
     if(parsedValue == null) return false;
 
-    List<String> parts = value.split('.');
+    List<String> parts = split('.');
     int integers = parts[0].length;
     int decimals = parts.length > 1? parts[1].length : 0;